backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
firebase-admin==6.5.0
google-generativeai==0.7.2
google-api-python-client==2.134.0
google-auth==2.30.0
pydantic==2.7.4
python-dotenv==1.0.1
